refactor(move): Extract logger into root package out of utils subpackage. Update imports where used.
