+++ tests/unit/conftest.py
@@ -17,8 +17,8 @@ from antlr4.error.ErrorListener import ErrorListener
 from fhy.lang.ast import ASTNode
 from fhy.lang.converter.from_parse_tree import from_parse_tree
 from fhy.lang.parser import FhYLexer, FhYParser
+from fhy.logger import get_logger
 from fhy.utils import error as fhy_error
-from fhy.utils.logger import get_logger
 
 log = get_logger(__name__, 10)
