+++ tests/unit/test_logger.py
